requirements.txt
matplotlib>=3.5.0
Pillow>=9.0.0
pandas>=1.3.0
fpdf2>=2.5.0
pyinstaller>=5.0.0